Refactor gamepad axis detection to base class
